pyproject.toml
[tool.poetry]
name = "scrobble-database"
version = "0.1.0"
description = ""
authors = ["Tamás Börcsök <tborcsok@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.5"
pandas = "^1.5.2"
python-dotenv = "^0.21.0"
requests = "^2.28.1"
tqdm = "^4.64.1"
psycopg2 = "^2.9.5"
click = "^8.1.3"
joblib = "^1.2.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.10.1"
mypy = "^0.991"
pylint = "^2.15.8"
ipykernel = "^6.18.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
